+++ lib/rubygems/credential_store/native/linux.rb
