+++ lib/rubygems/credential_store/native/macos.rb
