+++ lib/rubygems/credential_store/native/windows.rb
